+++ src/app/component/modal/modal.component.ts
@@ -1,23 +1,25 @@
 import {CommonModule} from '@angular/common';
 import {
-  NgModule, Component, OnInit, Input, Output, EventEmitter,
-  Renderer2, ElementRef, ViewChild, AfterContentInit, AfterViewInit, OnDestroy
+  NgModule, Component, Input, Output, EventEmitter,
+  Renderer2, ElementRef, ViewChild, AfterViewInit, OnDestroy, HostListener
 } from '@angular/core';
 import { trigger, style, state, animate, transition} from '@angular/animations';
 import { ShareModule } from '../common/share';
 import {ButtonModule} from '../button/button.directive';
 import {LoadingModule} from '../loading/loading.component';
+import {DomRenderer} from '../common/dom';
 
 @Component({
   selector: 'free-modal',
   template: `
-    <div #modal class="free-modal" [style.width.px]="width" [style.height.px]="height"
-         [@fadeInScale]="modalClass" (@fadeInScale.start)="animationEnd($event)"
-         [style.display]="visible ? 'block' : 'none'" [ngStyle]="{'left.px': left, 'top.px': top}">
+    <div *ngIf="_visible" class="free-modal"
+         [ngStyle]="{'z-index': zIndex, width: width + 'px',
+         height: height + 'px','left.px': left, 'top.px': top}"
+         [@fadeInScale]="modalState" [class.free-modal-spinner]="spinner">
       <div class="free-modal-header" *ngIf="!spinner" (mousedown)="onMouseDown($event)">
         <span *ngIf="header">{{header}}</span>
         <span><ng-content select="f-header"></ng-content></span>
-        <span *ngIf="closeIcon" class="free-modal-close" (click)="close()">
+        <span *ngIf="closeIcon" class="free-modal-close" (click)="visible = false">
           <i class="fa fa-close"></i>
         </span>
       </div>
@@ -52,19 +54,19 @@ import {LoadingModule} from '../loading/loading.component';
   `,
   animations: [
     trigger('fadeInScale', [
-      state('active', style({
+      state('in', style({
         opacity: 1,
-        transform: 'scale(1)'
+        transform: 'translate(-50%, -50%) scale(1)'
       })),
-      state('inactive', style({
-        opacity: 0.7,
-        transform: 'scale(0)'
-      })),
-      transition('active <=> inactive', animate('.3s ease'))
+      transition('void => *', [
+        style({opacity: 0.7, transform: 'translate(-50%, -50%) scale(0)'}),
+        animate(300, style({opacity: 1, transform: 'translate(-50%, -50%) scale(1)'}))
+      ])
     ])
-  ]
+  ],
+  providers: [DomRenderer]
 })
-export class ModalComponent implements AfterContentInit, AfterViewInit, OnDestroy {
+export class ModalComponent implements AfterViewInit, OnDestroy {
 
   @Input() header: string;
   @Input() width: any;
@@ -76,23 +78,34 @@ export class ModalComponent implements AfterContentInit, AfterViewInit, OnDestro
   @Input() spinner: string;
   @Output() visibleChange: EventEmitter<any> = new EventEmitter();
   @Output() onChange: EventEmitter<any> = new EventEmitter();
-  @ViewChild('modal') modalViewChild: ElementRef;
   @ViewChild('prompt') promptInput: ElementRef;
   _visible: boolean;
+  zIndex: number;
   modal: HTMLDivElement;
   mask: HTMLDivElement;
-  modalClass: string;
+  modalState: string;
   maskClickListener: Function;
-  container: any;
   isDown: boolean;
   point: any;
   left: number;
   top: number;
+  data: any;
+  initialized: boolean;
+  visibleChangeState: boolean;
+  documentClickListener: any;
   documentMousemoveListener: any;
   documentMouseupListener: any;
+  @HostListener('window:resize') onResize = () => {
+    this.center();
+  };
 
   constructor(public er: ElementRef,
-              public renderer2: Renderer2) { }
+              public domRenderer: DomRenderer,
+              public renderer2: Renderer2) {
+    this.data = {done: true};
+    this.modalState = 'in';
+    this.theme = 'default';
+  }
 
   @Input()
   get visible(): boolean {
@@ -101,38 +114,45 @@ export class ModalComponent implements AfterContentInit, AfterViewInit, OnDestro
 
   set visible(value: boolean) {
     this._visible = value;
-
-    if (this._visible) {
-      this.show();
-    } else {
-      this.close();
+    if (this.initialized) {
+      if (this._visible) {
+        this.show();
+      } else {
+        if (this.visibleChangeState) {
+          this.visibleChangeState = false;
+        } else {
+          this.close();
+        }
+      }
     }
   }
 
   ngAfterViewInit() {
-    this.modal = this.modalViewChild.nativeElement;
-    if (this.spinner) {
-      this.renderer2.addClass(this.modal, 'free-modal-spinner');
-    }
-  }
-
-  ngAfterContentInit() {
-    this.container = this.er.nativeElement;
-
-    const cancel = Array.from(this.container.querySelectorAll('.cancel'));
-    for (const c of cancel) {
-      this.renderer2.listen(c, 'click', () => this.close());
-    }
-  }
-
-  animationEnd(event: any) {
+    const modal = this.er.nativeElement;
+    this.documentClickListener = this.renderer2.listen(modal, 'click', (e) => {
+      if (this.initialized && this.visible) {
+        let target = e.target;
+        while (target) {
+          if (this.domRenderer.hasClass(target, 'free-modal')) {
+            break;
+          }
+          if (this.domRenderer.hasClass(target, 'cancel')) {
+            this.close();
+            break;
+          }
+          target = target.parentNode;
+        }
+      }
+    });
+    this.zIndex = 10002;
+    modal.classList.add('free-' + this.theme);
+    this.initialized = true;
   }
 
   show() {
+    this._visible = true;
     this.center();
-    this.modalClass = this.visible ? 'active' : 'inactive';
     this.addOverlay();
-
     if (this.delay) {
       setTimeout(() => {
         this.close();
@@ -141,21 +161,29 @@ export class ModalComponent implements AfterContentInit, AfterViewInit, OnDestro
   }
 
   confirm() {
-    let data = { value: null };
+    let data;
     if (this.type === 'prompt') {
-      data = { value: this.promptInput.nativeElement.value};
+      data = { value: this.promptInput.nativeElement.value, done: true};
+    } else {
+      data = {done: true};
     }
-    this.onChange.emit(data);
+    this.data = data;
     this.close();
   }
 
-  close() {
-    this.modalClass = this.visible ? 'active' : 'inactive';
-    this.visibleChange.emit(false);
+  hide() {
+    this.onChange.emit(this.data);
     if (this.mask) {
       this.renderer2.removeChild(document.body, this.mask);
     }
+  }
+
+  close() {
+    this.visibleChangeState = true;
+    this.hide();
+    this.visibleChange.emit(false);
     this.mask = null;
+    this.data = {done: true};
   }
 
   addOverlay() {
@@ -164,7 +192,7 @@ export class ModalComponent implements AfterContentInit, AfterViewInit, OnDestro
       this.mask.className = 'free-modal-mask';
       this.mask.style.cssText = 'position: fixed;top:0;left:0;width:100%;height:100%;' +
              'opacity:.5;background:#000;';
-      this.mask.style.zIndex = (parseInt(this.modal.style.zIndex, 10) - 1) + '';
+      this.mask.style.zIndex = (this.zIndex - 1) + '';
       this.maskClickListener = this.renderer2.listen(this.mask, 'click', (event: any) => {
         this.close();
       });
@@ -173,40 +201,19 @@ export class ModalComponent implements AfterContentInit, AfterViewInit, OnDestro
   }
 
   center() {
-    this.modal.style.zIndex = '10002';
-    this.modal.classList.add('free-' + this.theme);
-    const win = {
-      width: window.innerWidth,
-      height: window.innerHeight
-    };
-
-    let mw = {
-      width: this.modal.offsetWidth,
-      height: this.modal.offsetHeight
-    };
-
-    if (mw.width === 0 || mw.height === 0) {
-      this.modal.style.visibility = 'hidden';
-      this.modal.style.display = 'block';
-      mw = {
-        width: this.modal.offsetWidth,
-        height: this.modal.offsetHeight
-      };
-      this.modal.style.visibility = 'visible';
-      this.modal.style.display = 'none';
-    }
-
-    this.left = (win.width - mw.width) / 2;
-    this.top = (win.height - mw.height) / 2;
+    this.left = window.innerWidth / 2;
+    this.top = window.innerHeight / 2;
   }
 
   onMouseDown(event: any) {
+    this.modal = this.er.nativeElement.firstElementChild;
     this.isDown = true;
     this.point = {
       pageX: event.pageX,
       pageY: event.pageY
     };
     this.documentMousemoveListener = this.renderer2.listen('document', 'mousemove', (e) => {
+      this.modal.style.transitionDuration = '0ms';
       if (this.isDown) {
         this.left += e.pageX - this.point.pageX;
         this.top += e.pageY - this.point.pageY;
@@ -218,6 +225,7 @@ export class ModalComponent implements AfterContentInit, AfterViewInit, OnDestro
     });
     this.documentMousemoveListener = this.renderer2.listen('document', 'mouseup', (e) => {
       this.isDown = false;
+      this.modal.style.transitionDuration = '300ms';
     });
   }
 
@@ -233,7 +241,12 @@ export class ModalComponent implements AfterContentInit, AfterViewInit, OnDestro
   }
 
   ngOnDestroy() {
+    this.visibleChangeState = false;
     this.unbindDocumentMouseListener();
+    if (this.documentClickListener) {
+      this.documentClickListener();
+      this.documentClickListener = null;
+    }
   }
 }