+++ src/app/democase/main-fullpage/review-fullpage.component.ts
@@ -5,16 +5,16 @@ import {Component} from '@angular/core';
   template: `
     <free-fullpage [pagination]="true" [paginationClickable]="true">
       <free-fullpage-slide>
-        <div class="fullpage"><img src="freeng/assets/images/f1.jpg" alt=""></div>
+        <div class="fullpage"><img src="assets/images/f1.jpg" alt=""></div>
       </free-fullpage-slide>
       <free-fullpage-slide>
-        <div class="fullpage"><img src="freeng/assets/images/f2.jpg" alt=""></div>
+        <div class="fullpage"><img src="assets/images/f2.jpg" alt=""></div>
       </free-fullpage-slide>
       <free-fullpage-slide>
-        <div class="fullpage"><img src="freeng/assets/images/f3.jpg" alt=""></div>
+        <div class="fullpage"><img src="assets/images/f3.jpg" alt=""></div>
       </free-fullpage-slide>
       <free-fullpage-slide>
-        <div class="fullpage"><img src="freeng/assets/images/f4.jpg" alt=""></div>
+        <div class="fullpage"><img src="assets/images/f4.jpg" alt=""></div>
       </free-fullpage-slide>
     </free-fullpage>
   `,